inital version of spec v3 support
