+++ src/standards/v3.ts
@@ -1,0 +1,16 @@
+import { breaking, nonBreaking, unclassified } from '../constants';
+
+// TODO: check the new spec changes
+
+/**
+ * The standard object for AsyncAPI v3
+ *
+ * @private
+ */
+export const standard = {
+  '/asyncapi': {
+    add: nonBreaking,
+    remove: breaking,
+    edit: breaking,
+  },
+};
